Refactor driver update layout and libraries
