+++ src/heart/manage/driver_update/layout.py
@@ -1,0 +1,4 @@
+"""Driver file layout constants."""
+
+DRIVER_SETTINGS_FILENAME = "settings.toml"
+DRIVER_FILES = ("boot.py", "code.py", DRIVER_SETTINGS_FILENAME)
