+++ src/heart/manage/update.py
@@ -2,8 +2,8 @@ import sys
 
 from heart.manage.driver_update.configuration import load_driver_config
 from heart.manage.driver_update.exceptions import UpdateError
-from heart.manage.driver_update.filesystem import (DRIVER_SETTINGS_FILENAME,
-                                                   ensure_driver_files)
+from heart.manage.driver_update.filesystem import ensure_driver_files
+from heart.manage.driver_update.layout import DRIVER_SETTINGS_FILENAME
 from heart.manage.driver_update.mounts import (circuitpy_mounts,
                                                install_uf2_if_available,
                                                mount_points,
